+++ database/migrations/2026_02_07_162319_create_pacientes_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
 {
     Schema::create('pacientes', function (Blueprint $table) {
        $table->id();
-        $table->integer('cedula')->unique();
+        $table->bigInteger('cedula')->unique();
         $table->enum('categoria', ['estudiante', 'personal']);
         $table->timestamps();
